feature(get-json-from-file-table) mv out from client
